+++ core-domain/src/commonMain/kotlin/ai/saniou/thread/domain/usecase/thread/GetTopicCommentsPagerUseCase.kt
@@ -8,7 +8,19 @@ import kotlinx.coroutines.flow.Flow
 class GetTopicCommentsPagerUseCase(
     private val topicRepository: TopicRepository
 ) {
-    operator fun invoke(sourceId: String, topicId: String, isPoOnly: Boolean): Flow<PagingData<Comment>> {
-        return topicRepository.getTopicCommentsPager(sourceId, topicId, isPoOnly)
+    operator fun invoke(
+        sourceId: String,
+        topicId: String,
+        isPoOnly: Boolean,
+        isReverse: Boolean = false,
+        startPage: Int = 1,
+    ): Flow<PagingData<Comment>> {
+        return topicRepository.getTopicCommentsPager(
+            sourceId = sourceId,
+            topicId = topicId,
+            isPoOnly = isPoOnly,
+            isReverse = isReverse,
+            startPage = startPage,
+        )
     }
 }